Ensure `cuda_compiler_version` is set in tests
Previously the tests relied on `cuda_compiler_version` already being
set. However this is leftover from building the package where that
version is set. If one tests the package artifact in a different context
than the original build, this environment variable may not be set. To
fix this, add a step to set this environment variable before running the
tests to ensure it is available.